Reorganização
- Todas as views estão organizadas por tipo de usuário que acessa a página
- Os tipos de usuário foram padronizados para autarquia, instituicao_ensino e campo_estagio
- Removidas redundâncias em nomes de arquivos (com as mesmas palavras que a pasta que eles estão)
- Adicionada um plugin do eslint que verifica require()
- Adicionado uma configuração que recomenda extensões para se instalar no projeto
- views/partials renomeado para views/_partials para destacar entre outras páginas
- rotas que já existem mudaram de lugar
- a página inicial do hello world tem links para as outras páginas
