+++ .gitignore
@@ -113,4 +113,4 @@ dist
 
 
 # VS Code settings
-.vscode
+.vscode/settings.json
