--- controllers/autarquia-controller.js
@@ -1,11 +1,0 @@
-//const {Autarquia} = require("../models/autarquia-model")
-
-function renderCadLocEs(req, res) {
-    res.render("autarquia/cadLocalEstagio")
-}
-
-function renderCadIES(req, res) {
-    res.render("autarquia/cadIES")
-}
-
-module.exports = { renderCadLocEs, renderCadIES }
