+++ controllers/hello_world.js
@@ -3,7 +3,7 @@ const { Emoji } = require("../models/emoji")
 function renderHelloWorld(req, res) {
     const emoji = new Emoji()
     emoji.now(datetime => {
-        res.render("index.ejs", {
+        res.render("placeholder/hello_world.ejs", {
             mensagem: `Olá mundo! ${emoji.char}`,
             outramensagem: `${datetime}`
         })
