+++ package.json
@@ -3,6 +3,7 @@
     "body-parser": "^1.20.2",
     "dotenv": "^16.0.3",
     "ejs": "^3.1.9",
+    "eslint-plugin-import": "^2.27.5",
     "express": "^4.18.2",
     "mysql": "^2.18.1",
     "nodemon": "^2.0.22"
