+++ routes/campo_estagio.js
@@ -1,4 +1,4 @@
-const controller = require("../controllers/campodeestagio")
+const controller = require("../controllers/campo_estagio")
 const router = require("express").Router()
 
 router.post("/", controller.cadastrar)
