+++ routes/hello_world.js
@@ -1,4 +1,4 @@
-const controller = require("../controllers/helloworld")
+const controller = require("../controllers/hello_world")
 const router = require("express").Router()
 
 router.get("/", controller.renderHelloWorld)
