+++ views/base.ejs
@@ -2,7 +2,7 @@
 <html lang="pt-BR">
 
 <head>
-    <%- include(process.cwd() + '/views/partials/head.ejs') %>
+    <%- include(process.cwd() + '/views/_partials/head.ejs') %>
 </head>
 
 <body>
