+++ views/campo_estagio/atualizar-vagas.ejs
@@ -2,7 +2,7 @@
 <html lang="pt-BR">
 
 <head>
-    <%- include(process.cwd() + '/views/partials/head.ejs') %>
+    <%- include(process.cwd() + '/views/_partials/head.ejs') %>
 </head>
 
 <body>
@@ -17,7 +17,7 @@
                             <a href="dashboard.html" class="list-group-item">Cadastrar IES</a>
                         </li>
                         <li class="page_item page-item-2">
-                            <a href="profile.html" class="list-group-item ">Cadastro Local Estágio</a>
+                            <a href="profile.html" class="list-group-item ">Cadastro Campo de Estágio</a>
                         </li>
                         <li class="page_item page-item-2">
                             <a href="profile.html" class="list-group-item ">Solicitações de Estágio <span
