+++ views/login/login.ejs
@@ -2,8 +2,7 @@
 <html lang="pt-BR">
 
 <head>
-    <%- include(process.cwd() + '/views/partials/head.ejs') %>
-    <link rel="stylesheet" href="./login.css">
+    <%- include(process.cwd() + '/views/_partials/head.ejs') %>
 </head>
 
 <body>
