+++ views/placeholder/hello_world.ejs
@@ -19,17 +19,17 @@
     <div class="container mt-5">
         <div class="row">
             <div class="col-sm-4">
-                <h3>Column 1</h3>
+                <h3>Autarquia</h3>
+                <p><a href="./autarquia/cadastrar/campo_estagio">./autarquia/cadastrar/campo_estagio</a></p>
+                <p><a href="./autarquia/cadastrar/instituicao_ensino">./autarquia/cadastrar/instituicao_ensino</a></p>
+            </div>
+            <div class="col-sm-4">
+                <h3>Instiuição de Ensino</h3>
                 <p>Lorem ipsum dolor sit amet, consectetur adipisicing elit...</p>
                 <p>Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris...</p>
             </div>
             <div class="col-sm-4">
-                <h3>Column 2</h3>
-                <p>Lorem ipsum dolor sit amet, consectetur adipisicing elit...</p>
-                <p>Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris...</p>
-            </div>
-            <div class="col-sm-4">
-                <h3>Column 3</h3>
+                <h3>Campo de Estágio</h3>
                 <p>Lorem ipsum dolor sit amet, consectetur adipisicing elit...</p>
                 <p>Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris...</p>
             </div>
